remove unused commands
