+++ editor/src/components/canvas/commands/utils/property-utils.ts
@@ -1,0 +1,48 @@
+import type { ElementPath, JSXElement } from 'utopia-shared/src/types'
+import { foldEither } from '../../../../core/shared/either'
+import type { ValueAtPath } from '../../../../core/shared/jsx-attributes'
+import { setJSXValuesAtPaths } from '../../../../core/shared/jsx-attributes'
+import type { EditorState, EditorStatePatch } from '../../../editor/store/editor-state'
+import { modifyUnderlyingElementForOpenFile } from '../../../editor/store/editor-state'
+import { patchParseSuccessAtElementPath } from '../patch-utils'
+
+export function applyValuesAtPath(
+  editorState: EditorState,
+  target: ElementPath,
+  jsxValuesAndPathsToSet: ValueAtPath[],
+): { editorStateWithChanges: EditorState; editorStatePatch: EditorStatePatch } {
+  const workingEditorState = modifyUnderlyingElementForOpenFile(
+    target,
+    editorState,
+    (element: JSXElement) => {
+      return foldEither(
+        () => {
+          return element
+        },
+        (updatedProps) => {
+          return {
+            ...element,
+            props: updatedProps,
+          }
+        },
+        setJSXValuesAtPaths(element.props, jsxValuesAndPathsToSet),
+      )
+    },
+  )
+
+  const editorStatePatch = patchParseSuccessAtElementPath(target, workingEditorState, (success) => {
+    return {
+      topLevelElements: {
+        $set: success.topLevelElements,
+      },
+      imports: {
+        $set: success.imports,
+      },
+    }
+  })
+
+  return {
+    editorStateWithChanges: workingEditorState,
+    editorStatePatch: editorStatePatch,
+  }
+}
